Move gql strings to their own file, run black on test_gql
