+++ clients/python/test_gql.py
@@ -6,13 +6,13 @@ Will be replaced as soon as I work out a good solution for mocking the API"""
 
 
 def test_get_auth_token():
-    """ Check to see if get_auth_token returns a proper token.
+    """Check to see if get_auth_token returns a proper token.
     Will fail if credentials are not present in environmental variables.
     """
     assert re.match(
-        r"^[A-Za-z0-9-_=]+\.[A-Za-z0-9-_=]+\.?[A-Za-z0-9-_.+/=]*$",
-        get_auth_token()
+        r"^[A-Za-z0-9-_=]+\.[A-Za-z0-9-_=]+\.?[A-Za-z0-9-_.+/=]*$", get_auth_token()
     )
+
 
 def test_create_client():
     """Check that create_client creates a client.
